backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator>=2.0.0

# Database
sqlalchemy==2.0.25
aiomysql==0.2.0
pymysql==1.1.0
cryptography==42.0.0
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.6

# RAG & AI
langchain>=0.1.10
langchain-community>=0.0.25
langchain-google-genai>=0.0.9
google-generativeai>=0.4.0
chromadb>=0.4.22
sentence-transformers==2.3.1

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
unstructured

# Redis & Rate Limiting
redis==5.0.1
aioredis==2.0.1

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0
tenacity==8.2.3

# Monitoring & Logging
structlog==24.1.0
